Refactor into Mzinga.Viewer and Mzinga.Viewer.WPF
