+++ scripts/Build.ps1
@@ -21,7 +21,7 @@ if (Test-Path "$OutputRoot\$TargetOutputDirectory") {
 
 Write-Host "Build release..."
 
-dotnet msbuild $BuildArgs.Split() -restore -p:PublishDir="$RepoRoot\$OutputRoot\$TargetOutputDirectory" "$SolutionPath"
+dotnet msbuild $BuildArgs.Split() -restore -p:Configuration=Release -target:Publish -p:PublishDir="$RepoRoot\$OutputRoot\$TargetOutputDirectory" "$SolutionPath"
 Copy-Item "README.md" -Destination "$OutputRoot\$TargetOutputDirectory\ReadMe.txt"
 Copy-Item "LICENSE.md" -Destination "$OutputRoot\$TargetOutputDirectory\License.txt"
 Copy-Item "CHANGELOG.md" -Destination "$OutputRoot\$TargetOutputDirectory\Changelog.txt"
