+++ scripts/BuildLinux.ps1
@@ -3,6 +3,6 @@ param()
 [string] $Product = "Mzinga"
 [string] $Target = "Linux"
 
-& "$PSScriptRoot\Build.ps1" -Product $Product -Target $Target -BuildArgs "-p:Configuration=CliRelease -target:Publish -p:RuntimeIdentifier=linux-x64 -p:PublishSingleFile=true -p:PublishTrimmed=true"
+& "$PSScriptRoot\Build.ps1" -Product $Product -Target $Target -BuildArgs "-p:RuntimeIdentifier=linux-x64 -p:PublishSingleFile=true -p:PublishTrimmed=true"
 
 & "$PSScriptRoot\TarRelease.ps1" -Product $Product -Target $Target
