+++ scripts/BuildPortable.ps1
@@ -3,6 +3,6 @@ param()
 [string] $Product = "Mzinga"
 [string] $Target = "Portable"
 
-& "$PSScriptRoot\Build.ps1" -Product $Product -Target $Target -BuildArgs "-p:Configuration=Release -target:Publish"
+& "$PSScriptRoot\Build.ps1" -Product $Product -Target $Target
 
 & "$PSScriptRoot\ZipRelease.ps1" -Product $Product -Target $Target
