+++ scripts/BuildWindows.ps1
@@ -3,6 +3,6 @@ param()
 [string] $Product = "Mzinga"
 [string] $Target = "Windows"
 
-& "$PSScriptRoot\Build.ps1" -Product $Product -Target $Target -BuildArgs "-p:Configuration=Release -target:Publish -p:RuntimeIdentifier=win-x86 -p:PublishSingleFile=true -p:PublishTrimmed=true"
+& "$PSScriptRoot\Build.ps1" -Product $Product -Target $Target -BuildArgs "-p:RuntimeIdentifier=win-x86 -p:PublishSingleFile=true -p:PublishTrimmed=true"
 
 & "$PSScriptRoot\ZipRelease.ps1" -Product $Product -Target $Target
