+++ src/Mzinga.Test/Mzinga.Test.csproj
@@ -6,7 +6,6 @@
     <RootNamespace>Mzinga.Test</RootNamespace>
     <IsPackable>false</IsPackable>
     <IsPublishable>false</IsPublishable>
-    <Configurations>Debug;Release;CliRelease</Configurations>
   </PropertyGroup>
 
   <ItemGroup>
