+++ src/Mzinga.Trainer/Mzinga.Trainer.csproj
@@ -6,7 +6,6 @@
     <StartupObject>Mzinga.Trainer.Program</StartupObject>
     <AssemblyName>Mzinga.Trainer</AssemblyName>
     <RootNamespace>Mzinga.Trainer</RootNamespace>
-    <Configurations>Debug;Release;CliRelease</Configurations>
   </PropertyGroup>
 
   <Import Project="..\Mzinga.Shared\Mzinga.Shared.projitems" Label="Shared" />
