+++ src/Mzinga.Viewer.WPF/App.xaml
