+++ src/Mzinga.Viewer.WPF/App.xaml.cs
