+++ src/Mzinga.Viewer.WPF/EngineConsoleWindow.xaml
