+++ src/Mzinga.Viewer.WPF/ExceptionWindow.xaml
