+++ src/Mzinga.Viewer.WPF/GameMetadataWindow.xaml
