+++ src/Mzinga.Viewer.WPF/InformationWindow.xaml
