+++ src/Mzinga.Viewer.WPF/LicensesWindow.xaml.cs
