+++ src/Mzinga.Viewer.WPF/MainWindow.xaml.cs
