+++ src/Mzinga.Viewer.WPF/NewGameWindow.xaml
