+++ src/Mzinga.Viewer.WPF/NewGameWindow.xaml.cs
