+++ src/Mzinga.Viewer.WPF/Program.cs
@@ -1,10 +1,10 @@
-// 
+﻿// 
 // Program.cs
 //  
 // Author:
 //       Jon Thysell <thysell@gmail.com>
 // 
-// Copyright (c) 2021 Jon Thysell <http://jonthysell.com>
+// Copyright (c) 2018, 2021 Jon Thysell <http://jonthysell.com>
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -25,35 +25,33 @@
 // THE SOFTWARE.
 
 using System;
-using System.Diagnostics;
-
-using Avalonia;
+using System.Windows;
 
 namespace Mzinga.Viewer
 {
-    class Program
+    public class Program
     {
-        // Initialization code. Don't use any Avalonia, third-party APIs or any
-        // SynchronizationContext-reliant code before AppMain is called: things aren't initialized
-        // yet and stuff might break.
+        private static string configFile = null;
+
+        [STAThread]
         public static void Main(string[] args)
         {
-            AppDomain.CurrentDomain.UnhandledException += CurrentDomain_UnhandledException;
-            BuildAvaloniaApp().StartWithClassicDesktopLifetime(args);
-        }
+            try
+            {
+                if (null != args && args.Length > 0)
+                {
+                    configFile = args[0];
+                }
 
-        // Avalonia configuration, don't remove; also used by visual designer.
-        public static AppBuilder BuildAvaloniaApp()
-        {
-            return AppBuilder.Configure<App>()
-                .UsePlatformDetect()
-                .LogToTrace()
-                .With(new AvaloniaNativePlatformOptions { UseGpu = false });
-        }
-
-        private static void CurrentDomain_UnhandledException(object sender, UnhandledExceptionEventArgs e)
-        {
-            Trace.TraceError($"Unhandled Exception: { (e.ExceptionObject as Exception)?.ToString() }");
+                App app = new App(configFile);
+                app.InitializeComponent();
+                app.Run();
+            }
+            catch (Exception ex)
+            {
+                string message = string.Join(Environment.NewLine, string.Format("Unhandled Error: {0}", ex.Message), ex.StackTrace);
+                MessageBox.Show(message, "Unhandled Error", MessageBoxButton.OK, MessageBoxImage.Error);
+            }
         }
     }
 }
