+++ src/Mzinga.Viewer.WPF/Resources/mzinga.ico
