+++ src/Mzinga.Viewer.WPF/ScrollingTextBox.cs
