+++ src/Mzinga.Viewer.WPF/ViewerConfigWindow.xaml.cs
