+++ src/Mzinga.Viewer/MessageHandlers.cs
@@ -4,7 +4,7 @@
 // Author:
 //       Jon Thysell <thysell@gmail.com>
 // 
-// Copyright (c) 2015, 2016, 2017, 2018, 2019 Jon Thysell <http://jonthysell.com>
+// Copyright (c) 2015, 2016, 2017, 2018, 2019, 2021 Jon Thysell <http://jonthysell.com>
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -25,11 +25,14 @@
 // THE SOFTWARE.
 
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Windows;
+using System.Threading.Tasks;
 
-using Microsoft.Win32;
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Controls.ApplicationLifetimes;
 
 using GalaSoft.MvvmLight.Messaging;
 
@@ -40,16 +43,18 @@ namespace Mzinga.Viewer
 {
     public class MessageHandlers
     {
+        public static Window MainWindow => (Application.Current.ApplicationLifetime as IClassicDesktopStyleApplicationLifetime).MainWindow;
+
         public static void RegisterMessageHandlers(object recipient)
         {
             Messenger.Default.Register<ExceptionMessage>(recipient, (message) => ShowException(message));
             Messenger.Default.Register<InformationMessage>(recipient, (message) => ShowInformation(message));
             Messenger.Default.Register<ConfirmationMessage>(recipient, (message) => ShowConfirmation(message));
-            Messenger.Default.Register<LaunchUrlMessage>(recipient, (message) => LaunchUrl(message));
+            Messenger.Default.Register<LaunchUrlMessage>(recipient, async (message) => await LaunchUrlAsync(message));
             Messenger.Default.Register<ShowLicensesMessage>(recipient, (message) => ShowLicense(message));
-            Messenger.Default.Register<NewGameMessage>(recipient, (message) => ShowNewGame(message));
-            Messenger.Default.Register<LoadGameMessage>(recipient, (message) => ShowLoadGame(message));
-            Messenger.Default.Register<SaveGameMessage>(recipient, (message) => ShowSaveGame(message));
+            Messenger.Default.Register<NewGameMessage>(recipient, async (message) => await ShowNewGameAsync(message));
+            Messenger.Default.Register<LoadGameMessage>(recipient, async (message) => await ShowLoadGameAsync(message));
+            Messenger.Default.Register<SaveGameMessage>(recipient, async (message) => await ShowSaveGameAsync(message));
             Messenger.Default.Register<GameMetadataMessage>(recipient, (message) => ShowGameMetadata(message));
             Messenger.Default.Register<ViewerConfigMessage>(recipient, (message) => ShowViewerConfig(message));
             Messenger.Default.Register<EngineOptionsMessage>(recipient, (message) => ShowEngineOptions(message));
@@ -74,36 +79,36 @@ namespace Mzinga.Viewer
 
         private static void ShowException(ExceptionMessage message)
         {
-            ExceptionWindow window = new ExceptionWindow
-            {
-                DataContext = message.ExceptionVM,
-                Owner = Application.Current.MainWindow,
-            };
-            message.ExceptionVM.RequestClose += (sender, e) =>
-            {
-                window.Close();
-            };
-            window.ShowDialog();
+            //ExceptionWindow window = new ExceptionWindow
+            //{
+            //    DataContext = message.ExceptionVM,
+            //    Owner = Application.Current.MainWindow,
+            //};
+            //message.ExceptionVM.RequestClose += (sender, e) =>
+            //{
+            //    window.Close();
+            //};
+            //window.ShowDialog();
         }
 
         private static void ShowInformation(InformationMessage message)
         {
             try
             {
-                InformationWindow window = new InformationWindow
-                {
-                    DataContext = message.InformationVM,
-                    Owner = Application.Current.MainWindow,
-                };
-                message.InformationVM.RequestClose += (sender, e) =>
-                {
-                    window.Close();
-                };
-                window.Closed += (sender, args) =>
-                {
-                    message.Process();
-                };
-                window.Show();
+                //InformationWindow window = new InformationWindow
+                //{
+                //    DataContext = message.InformationVM,
+                //    Owner = Application.Current.MainWindow,
+                //};
+                //message.InformationVM.RequestClose += (sender, e) =>
+                //{
+                //    window.Close();
+                //};
+                //window.Closed += (sender, args) =>
+                //{
+                //    message.Process();
+                //};
+                //window.Show();
             }
             catch (Exception ex)
             {
@@ -115,8 +120,9 @@ namespace Mzinga.Viewer
         {
             try
             {
-                MessageBoxResult result = MessageBox.Show(message.Message, "Mzinga", MessageBoxButton.YesNo);
-                message.Process(result == MessageBoxResult.Yes);
+                //MessageBoxResult result = MessageBox.Show(message.Message, "Mzinga", MessageBoxButton.YesNo);
+                //message.Process(result == MessageBoxResult.Yes);
+                message.Process(true);
             }
             catch (Exception ex)
             {
@@ -124,75 +130,99 @@ namespace Mzinga.Viewer
             }
         }
 
-        private static void LaunchUrl(LaunchUrlMessage message)
+        private static async Task LaunchUrlAsync(LaunchUrlMessage message)
         {
-            try
+            string url = message.Url;
+
+            await Task.Run(() =>
             {
-                Process.Start(message.Url, null);
-            }
-            catch (Exception ex)
-            {
-                ExceptionUtils.HandleException(ex);
-            }
+                try
+                {
+                    if (AppInfo.IsWindows)
+                    {
+                        Process.Start(new ProcessStartInfo(url)
+                        {
+                            UseShellExecute = true
+                        });
+                    }
+                    else if (AppInfo.IsMacOS)
+                    {
+                        Process.Start("open", url);
+                    }
+                    else if (AppInfo.IsLinux)
+                    {
+                        Process.Start("xdg-open", url);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    ExceptionUtils.HandleException(ex);
+                }
+            });
         }
 
         private static void ShowLicense(ShowLicensesMessage message)
         {
-            LicensesWindow window = new LicensesWindow()
-            {
-                DataContext = message.LicensesVM,
-                Owner = Application.Current.MainWindow
-            };
+            //LicensesWindow window = new LicensesWindow()
+            //{
+            //    DataContext = message.LicensesVM,
+            //    Owner = Application.Current.MainWindow
+            //};
 
-            message.LicensesVM.Licenses.Add(GetExtendedWpfToolkitLicense());
-            message.LicensesVM.RequestClose += () =>
-            {
-                window.Close();
-            };
-            window.ShowDialog();
+            //message.LicensesVM.Licenses.Add(GetExtendedWpfToolkitLicense());
+            //message.LicensesVM.RequestClose += () =>
+            //{
+            //    window.Close();
+            //};
+            //window.ShowDialog();
             message.Process();
         }
 
-        private static void ShowNewGame(NewGameMessage message)
+        private static async Task ShowNewGameAsync(NewGameMessage message)
         {
-            try
+            await Task.Run(() =>
             {
-                NewGameWindow window = new NewGameWindow
+                try
                 {
-                    DataContext = message.NewGameVM,
-                    Owner = Application.Current.MainWindow,
-                };
-                message.NewGameVM.RequestClose += (sender, e) =>
+                    //NewGameWindow window = new NewGameWindow
+                    //{
+                    //    DataContext = message.NewGameVM,
+                    //    Owner = Application.Current.MainWindow,
+                    //};
+                    //message.NewGameVM.RequestClose += (sender, e) =>
+                    //{
+                    //    window.Close();
+                    //};
+                    //window.ShowDialog();
+                    message.NewGameVM.Accept.Execute(null);
+                    message.Process();
+                }
+                catch (Exception ex)
                 {
-                    window.Close();
-                };
-                window.ShowDialog();
-                message.Process();
-            }
-            catch (Exception ex)
-            {
-                ExceptionUtils.HandleException(ex);
-            }
+                    ExceptionUtils.HandleException(ex);
+                }
+            });
         }
 
-        private static void ShowLoadGame(LoadGameMessage message)
+        private static async Task ShowLoadGameAsync(LoadGameMessage message)
         {
             GameRecording gr = null;
             try
             {
                 OpenFileDialog dialog = new OpenFileDialog
                 {
+                    AllowMultiple = false,
                     Title = "Open Game",
-                    DefaultExt = ".pgn",
-                    Filter = "All Supported Files|*.pgn;*.sgf|Portable Game Notation|*.pgn|BoardSpace Smart Game Format|*.sgf",
-                    AddExtension = true
+                    Filters = GetFilters(true),
                 };
 
-                if (dialog.ShowDialog(Application.Current.MainWindow).GetValueOrDefault())
+                string[] filenames = await dialog.ShowAsync(MainWindow);
+
+                if (null != filenames && filenames.Length > 0 && !string.IsNullOrWhiteSpace(filenames[0]))
                 {
-                    using (Stream inputStream = dialog.OpenFile())
+                    string fileName = filenames[0].Trim();
+                    using (Stream inputStream = File.OpenRead(fileName))
                     {
-                        string fileName = dialog.SafeFileName;
                         gr = Path.GetExtension(fileName).ToLower() == ".sgf" ? GameRecording.LoadSGF(inputStream, fileName) : GameRecording.LoadPGN(inputStream, fileName);
                     }
                 }
@@ -207,7 +237,7 @@ namespace Mzinga.Viewer
             }
         }
 
-        private static void ShowSaveGame(SaveGameMessage message)
+        private static async Task ShowSaveGameAsync(SaveGameMessage message)
         {
             string fileName = null;
             try
@@ -215,18 +245,21 @@ namespace Mzinga.Viewer
                 SaveFileDialog dialog = new SaveFileDialog
                 {
                     Title = "Save Game",
-                    DefaultExt = ".pgn",
-                    Filter = "Portable Game Notation|*.pgn",
-                    AddExtension = true,
-                    FileName = Path.GetFileNameWithoutExtension(message.GameRecording.FileName)
+                    DefaultExtension = ".pgn",
+                    Filters = GetFilters(false),
+                    Directory = !string.IsNullOrEmpty(message.GameRecording.FileName) ? Path.GetDirectoryName(message.GameRecording.FileName) : null,
+                    InitialFileName = !string.IsNullOrEmpty(message.GameRecording.FileName) ? Path.GetFileName(message.GameRecording.FileName) : null,
                 };
 
-                if (dialog.ShowDialog(Application.Current.MainWindow).GetValueOrDefault())
+                string result = await dialog.ShowAsync(MainWindow);
+
+                fileName = result?.Trim();
+
+                if (!string.IsNullOrEmpty(fileName))
                 {
-                    using (Stream outputStream = dialog.OpenFile())
+                    using (Stream outputStream = File.OpenWrite(fileName))
                     {
                         message.GameRecording.SavePGN(outputStream);
-                        fileName = dialog.SafeFileName;
                     }
                 }
             }
@@ -240,21 +273,52 @@ namespace Mzinga.Viewer
             }
         }
 
+        private static List<FileDialogFilter> GetFilters(bool open)
+        {
+            var filters = new List<FileDialogFilter>();
+
+            if (open)
+            {
+                filters.Add(new FileDialogFilter()
+                {
+                    Name = "All Supported Files",
+                    Extensions = new List<string>() { "pgn", "sgf" }
+                });
+            }
+
+            filters.Add(new FileDialogFilter()
+            {
+                Name = "Portable Game Notation",
+                Extensions = new List<string>() { "pgn" }
+            });
+
+            if (open)
+            {
+                filters.Add(new FileDialogFilter()
+                {
+                    Name = "BoardSpace Smart Game Format",
+                    Extensions = new List<string>() { "sgf" }
+                });
+            }
+
+            return filters;
+        }
+
         private static void ShowGameMetadata(GameMetadataMessage message)
         {
             try
             {
-                GameMetadataWindow window = new GameMetadataWindow
-                {
-                    DataContext = message.GameMetadataVM,
-                    Owner = Application.Current.MainWindow,
-                };
-                message.GameMetadataVM.RequestClose += (sender, e) =>
-                {
-                    window.Close();
-                };
-                window.ShowDialog();
-                message.Process();
+                //GameMetadataWindow window = new GameMetadataWindow
+                //{
+                //    DataContext = message.GameMetadataVM,
+                //    Owner = Application.Current.MainWindow,
+                //};
+                //message.GameMetadataVM.RequestClose += (sender, e) =>
+                //{
+                //    window.Close();
+                //};
+                //window.ShowDialog();
+                //message.Process();
             }
             catch (Exception ex)
             {
@@ -266,17 +330,17 @@ namespace Mzinga.Viewer
         {
             try
             {
-                ViewerConfigWindow window = new ViewerConfigWindow
-                {
-                    DataContext = message.ViewerConfigVM,
-                    Owner = Application.Current.MainWindow,
-                };
-                message.ViewerConfigVM.RequestClose += (sender, e) =>
-                {
-                    window.Close();
-                };
-                window.ShowDialog();
-                message.Process();
+                //ViewerConfigWindow window = new ViewerConfigWindow
+                //{
+                //    DataContext = message.ViewerConfigVM,
+                //    Owner = Application.Current.MainWindow,
+                //};
+                //message.ViewerConfigVM.RequestClose += (sender, e) =>
+                //{
+                //    window.Close();
+                //};
+                //window.ShowDialog();
+                //message.Process();
             }
             catch (Exception ex)
             {
@@ -288,22 +352,22 @@ namespace Mzinga.Viewer
         {
             try
             {
-                if (message.EngineOptionsVM.Options.Count == 0)
-                {
-                    throw new Exception("This engine exposes no options to set.");
-                }
+                //if (message.EngineOptionsVM.Options.Count == 0)
+                //{
+                //    throw new Exception("This engine exposes no options to set.");
+                //}
 
-                EngineOptionsWindow window = new EngineOptionsWindow
-                {
-                    DataContext = message.EngineOptionsVM,
-                    Owner = Application.Current.MainWindow,
-                };
-                message.EngineOptionsVM.RequestClose += (sender, e) =>
-                {
-                    window.Close();
-                };
-                window.ShowDialog();
-                message.Process();
+                //EngineOptionsWindow window = new EngineOptionsWindow
+                //{
+                //    DataContext = message.EngineOptionsVM,
+                //    Owner = Application.Current.MainWindow,
+                //};
+                //message.EngineOptionsVM.RequestClose += (sender, e) =>
+                //{
+                //    window.Close();
+                //};
+                //window.ShowDialog();
+                //message.Process();
             }
             catch (Exception ex)
             {
@@ -315,42 +379,19 @@ namespace Mzinga.Viewer
         {
             try
             {
-                EngineConsoleWindow window = EngineConsoleWindow.Instance;
+                //EngineConsoleWindow window = EngineConsoleWindow.Instance;
 
-                window.Show();
+                //window.Show();
 
-                if (!window.IsActive)
-                {
-                    window.Activate();
-                }
+                //if (!window.IsActive)
+                //{
+                //    window.Activate();
+                //}
             }
             catch (Exception ex)
             {
                 ExceptionUtils.HandleException(ex);
             }
         }
-
-        private static ObservableLicense GetExtendedWpfToolkitLicense()
-        {
-            return new ObservableLicense("Extended WPF Toolkit", "Copyright © 2010-2019 Xceed Software Inc", "Microsoft Public License (Ms-PL)", string.Join(Environment.NewLine + Environment.NewLine, _msPlLicense));
-        }
-
-        private static readonly string[] _msPlLicense = {
-            @"This license governs use of the accompanying software. If you use the software, you accept this license. If you do not accept the license, do not use the software.",
-            @"1. Definitions",
-            @"The terms ""reproduce,"" ""reproduction,"" ""derivative works,"" and ""distribution"" have the same meaning here as under U.S. copyright law.",
-            @"A ""contribution"" is the original software, or any additions or changes to the software.",
-            @"A ""contributor"" is any person that distributes its contribution under this license.",
-            @"""Licensed patents"" are a contributor's patent claims that read directly on its contribution.",
-            @"2. Grant of Rights",
-            @"(A) Copyright Grant- Subject to the terms of this license, including the license conditions and limitations in section 3, each contributor grants you a non-exclusive, worldwide, royalty-free copyright license to reproduce its contribution, prepare derivative works of its contribution, and distribute its contribution or any derivative works that you create.",
-            @"(B) Patent Grant- Subject to the terms of this license, including the license conditions and limitations in section 3, each contributor grants you a non-exclusive, worldwide, royalty-free license under its licensed patents to make, have made, use, sell, offer for sale, import, and/or otherwise dispose of its contribution in the software or derivative works of the contribution in the software.",
-            @"3. Conditions and Limitations",
-            @"(A) No Trademark License- This license does not grant you rights to use any contributors' name, logo, or trademarks.",
-            @"(B) If you bring a patent claim against any contributor over patents that you claim are infringed by the software, your patent license from such contributor to the software ends automatically.",
-            @"(C) If you distribute any portion of the software, you must retain all copyright, patent, trademark, and attribution notices that are present in the software.",
-            @"(D) If you distribute any portion of the software in source code form, you may do so only under this license by including a complete copy of this license with your distribution. If you distribute any portion of the software in compiled or object code form, you may only do so under a license that complies with this license.",
-            @"(E) The software is licensed ""as-is."" You bear the risk of using it. The contributors give no express warranties, guarantees or conditions.You may have additional consumer rights under your local laws which this license cannot change.To the extent permitted under your local laws, the contributors exclude the implied warranties of merchantability, fitness for a particular purpose and non-infringement.",
-        };
     }
 }
