+++ src/Mzinga.Viewer/Resources/icon.icns
