+++ src/Mzinga.Viewer/Resources/icon.ico
