+++ src/Mzinga.sln
@@ -13,66 +13,54 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mzinga.Test", "Mzinga.Test\
 EndProject
 Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Mzinga.SharedUX", "Mzinga.SharedUX\Mzinga.SharedUX.shproj", "{766F35B9-DCFC-4358-8A3A-68E61C6D3BAA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mzinga.Viewer", "Mzinga.Viewer\Mzinga.Viewer.csproj", "{E3448B5E-41C3-4828-91F5-06BD49C73E22}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mzinga.Viewer.WPF", "Mzinga.Viewer.WPF\Mzinga.Viewer.WPF.csproj", "{E3448B5E-41C3-4828-91F5-06BD49C73E22}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mzinga.Perft", "Mzinga.Perft\Mzinga.Perft.csproj", "{BD3A0E3A-EB2A-409B-AF6E-F88FFBF34CC4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mzinga", "Mzinga\Mzinga.csproj", "{531FFFF0-0603-4EDA-8AB7-A780BA7B4908}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mzinga.Viewer", "Mzinga.Viewer\Mzinga.Viewer.csproj", "{A6BF3A8F-1B7A-47C6-A0D7-D32D75FF4099}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		Mzinga.Shared\Mzinga.Shared.projitems*{23c4244b-2965-41fb-9f35-bdcae2ab2c7b}*SharedItemsImports = 5
 		Mzinga.Shared\Mzinga.Shared.projitems*{36cc70c4-9bce-4546-8c4f-b116d5086291}*SharedItemsImports = 5
 		Mzinga.Shared\Mzinga.Shared.projitems*{3a664979-a4db-4c05-bd8f-c6a1799624df}*SharedItemsImports = 13
-		Mzinga.SharedUX\Mzinga.SharedUX.projitems*{531ffff0-0603-4eda-8ab7-a780ba7b4908}*SharedItemsImports = 5
-		Mzinga.Shared\Mzinga.Shared.projitems*{531ffff0-0603-4eda-8ab7-a780ba7b4908}*SharedItemsImports = 5
 		Mzinga.SharedUX\Mzinga.SharedUX.projitems*{766f35b9-dcfc-4358-8a3a-68e61c6d3baa}*SharedItemsImports = 13
+		Mzinga.SharedUX\Mzinga.SharedUX.projitems*{a6bf3a8f-1b7a-47c6-a0d7-d32d75ff4099}*SharedItemsImports = 5
+		Mzinga.Shared\Mzinga.Shared.projitems*{a6bf3a8f-1b7a-47c6-a0d7-d32d75ff4099}*SharedItemsImports = 5
 		Mzinga.Shared\Mzinga.Shared.projitems*{ada152d0-9a37-4884-a3b0-61fa7790bd9c}*SharedItemsImports = 5
 		Mzinga.Shared\Mzinga.Shared.projitems*{bd3a0e3a-eb2a-409b-af6e-f88ffbf34cc4}*SharedItemsImports = 5
 		Mzinga.SharedUX\Mzinga.SharedUX.projitems*{e3448b5e-41c3-4828-91f5-06bd49c73e22}*SharedItemsImports = 5
 		Mzinga.Shared\Mzinga.Shared.projitems*{e3448b5e-41c3-4828-91f5-06bd49c73e22}*SharedItemsImports = 5
 	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		CliRelease|Any CPU = CliRelease|Any CPU
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{23C4244B-2965-41FB-9F35-BDCAE2AB2C7B}.CliRelease|Any CPU.ActiveCfg = CliRelease|Any CPU
-		{23C4244B-2965-41FB-9F35-BDCAE2AB2C7B}.CliRelease|Any CPU.Build.0 = CliRelease|Any CPU
 		{23C4244B-2965-41FB-9F35-BDCAE2AB2C7B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{23C4244B-2965-41FB-9F35-BDCAE2AB2C7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{23C4244B-2965-41FB-9F35-BDCAE2AB2C7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{23C4244B-2965-41FB-9F35-BDCAE2AB2C7B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{36CC70C4-9BCE-4546-8C4F-B116D5086291}.CliRelease|Any CPU.ActiveCfg = CliRelease|Any CPU
-		{36CC70C4-9BCE-4546-8C4F-B116D5086291}.CliRelease|Any CPU.Build.0 = CliRelease|Any CPU
 		{36CC70C4-9BCE-4546-8C4F-B116D5086291}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{36CC70C4-9BCE-4546-8C4F-B116D5086291}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{36CC70C4-9BCE-4546-8C4F-B116D5086291}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{36CC70C4-9BCE-4546-8C4F-B116D5086291}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ADA152D0-9A37-4884-A3B0-61FA7790BD9C}.CliRelease|Any CPU.ActiveCfg = CliRelease|Any CPU
-		{ADA152D0-9A37-4884-A3B0-61FA7790BD9C}.CliRelease|Any CPU.Build.0 = CliRelease|Any CPU
 		{ADA152D0-9A37-4884-A3B0-61FA7790BD9C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{ADA152D0-9A37-4884-A3B0-61FA7790BD9C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ADA152D0-9A37-4884-A3B0-61FA7790BD9C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ADA152D0-9A37-4884-A3B0-61FA7790BD9C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E3448B5E-41C3-4828-91F5-06BD49C73E22}.CliRelease|Any CPU.ActiveCfg = Release|Any CPU
 		{E3448B5E-41C3-4828-91F5-06BD49C73E22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E3448B5E-41C3-4828-91F5-06BD49C73E22}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E3448B5E-41C3-4828-91F5-06BD49C73E22}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E3448B5E-41C3-4828-91F5-06BD49C73E22}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BD3A0E3A-EB2A-409B-AF6E-F88FFBF34CC4}.CliRelease|Any CPU.ActiveCfg = CliRelease|Any CPU
-		{BD3A0E3A-EB2A-409B-AF6E-F88FFBF34CC4}.CliRelease|Any CPU.Build.0 = CliRelease|Any CPU
 		{BD3A0E3A-EB2A-409B-AF6E-F88FFBF34CC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BD3A0E3A-EB2A-409B-AF6E-F88FFBF34CC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BD3A0E3A-EB2A-409B-AF6E-F88FFBF34CC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BD3A0E3A-EB2A-409B-AF6E-F88FFBF34CC4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{531FFFF0-0603-4EDA-8AB7-A780BA7B4908}.CliRelease|Any CPU.ActiveCfg = Release|Any CPU
-		{531FFFF0-0603-4EDA-8AB7-A780BA7B4908}.CliRelease|Any CPU.Build.0 = Release|Any CPU
-		{531FFFF0-0603-4EDA-8AB7-A780BA7B4908}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{531FFFF0-0603-4EDA-8AB7-A780BA7B4908}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{531FFFF0-0603-4EDA-8AB7-A780BA7B4908}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{531FFFF0-0603-4EDA-8AB7-A780BA7B4908}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A6BF3A8F-1B7A-47C6-A0D7-D32D75FF4099}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A6BF3A8F-1B7A-47C6-A0D7-D32D75FF4099}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A6BF3A8F-1B7A-47C6-A0D7-D32D75FF4099}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A6BF3A8F-1B7A-47C6-A0D7-D32D75FF4099}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
