--- src/Mzinga/Mzinga.csproj
@@ -1,40 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <StartupObject>Mzinga.Program</StartupObject>
-    <AssemblyName>Mzinga</AssemblyName>
-    <RootNamespace>Mzinga</RootNamespace>
-    <ApplicationIcon>Resources\icon.ico</ApplicationIcon>
-    <DefineConstants>TRACE;AVALONIAUI</DefineConstants>
-    <Configurations>Debug;Release</Configurations>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(RuntimeIdentifier)' == 'osx-x64' ">
-    <CFBundleName>Mzinga</CFBundleName>
-    <CFBundleDisplayName>Mzinga</CFBundleDisplayName>
-    <CFBundleIdentifier>com.jonthysell.mzinga</CFBundleIdentifier>
-    <CFBundlePackageType>AAPL</CFBundlePackageType>
-    <CFBundleExecutable>Mzinga</CFBundleExecutable>
-    <CFBundleIconFile>icon.icns</CFBundleIconFile>
-    <NSPrincipalClass>NSApplication</NSPrincipalClass>
-    <NSHighResolutionCapable>true</NSHighResolutionCapable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <AvaloniaResource Include="Resources\*.*" />
-  </ItemGroup>
-
-  <Import Project="..\Mzinga.SharedUX\Mzinga.SharedUX.projitems" Label="Shared" />
-
-  <Import Project="..\Mzinga.Shared\Mzinga.Shared.projitems" Label="Shared" />
-
-  <ItemGroup>
-    <PackageReference Include="Avalonia" Version="0.10.0" />
-    <PackageReference Include="Avalonia.Desktop" Version="0.10.0" />
-    <PackageReference Include="Avalonia.Diagnostics" Version="0.10.0" />
-    <PackageReference Include="Dotnet.Bundle" Version="0.9.13" />
-    <PackageReference Include="MvvmLightLibsStd10" Version="5.4.1.1" />
-   </ItemGroup>
-</Project>
